Map layer dataprovider descriptions.
