+++ control-base/src/main/java/fi/nls/oskari/control/layer/DataProviderHandler.java
@@ -4,7 +4,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import javax.servlet.http.HttpServletRequest;
+
 import fi.nls.oskari.cache.CacheManager;
+
+import org.json.JSONException;
+import org.json.JSONObject;
 import org.oskari.log.AuditLog;
 import org.oskari.service.util.ServiceFactory;
 
@@ -18,8 +23,13 @@ import fi.nls.oskari.domain.map.DataProvider;
 import fi.nls.oskari.domain.map.OskariLayer;
 import fi.nls.oskari.map.layer.DataProviderService;
 import fi.nls.oskari.map.layer.OskariLayerService;
+import fi.nls.oskari.util.ConversionHelper;
+import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
+import fi.nls.oskari.util.RequestHelper;
 import fi.nls.oskari.util.ResponseHelper;
+
+import static fi.nls.oskari.control.ActionConstants.PARAM_NAME_PREFIX;
 
 /**
  * Action route handler for single data provider related operations
@@ -105,4 +115,78 @@ public class DataProviderHandler extends RestActionHandler {
 	private void flushLayerListCache() {
 		CacheManager.getCache(GetMapLayerGroupsHandler.CACHE_NAME).flush(true);
 	}
+
+	public void handlePut(ActionParameters params) throws ActionException {
+        handlePost(params);
+    }
+    @Override
+    public void handlePost(ActionParameters params) throws ActionException {
+
+        final HttpServletRequest request = params.getRequest();
+        final DataProvider dataProvider = new DataProvider();
+        // hierarchical admin frontend sends separate params
+        dataProvider.setId(params.getHttpParam(PARAM_ID, -1));
+        dataProvider.setNames(RequestHelper.parsePrefixedParamsMap(request, PARAM_NAME_PREFIX));
+
+
+        if (dataProvider.getLocale() == null) {
+            // the classical layer admin sends JSON payload
+            // locale as payload  { id: "124", locales: {en: "name en", fi: "name fi"}}
+			try {
+				JSONObject payload = params.getPayLoadJSON();
+				dataProvider.setId(ConversionHelper.getInt(payload.optString("id"), -1));
+				JSONObject locales;
+				locales = payload.getJSONObject("locales");
+
+				JSONObject defaultLang = locales.optJSONObject(PropertyUtil.getDefaultLanguage());
+				if (defaultLang == null) {
+					throw new ActionParamsException("No locale for default lang: " + PropertyUtil.getDefaultLanguage());
+				}
+				String name = defaultLang.optString("name");
+				if (name == null || name.trim().isEmpty()) {
+					throw new ActionParamsException("No name for default lang: " + PropertyUtil.getDefaultLanguage());
+				}
+
+				dataProvider.setLocale(locales);
+			} catch (JSONException e) {
+				throw new ActionException("Cannot poulate maplayer group from request", e);
+			}
+        }
+
+        if (dataProvider.getLocale() == null) {
+            throw new ActionParamsException("Missing names for layer dataprovider!");
+        }
+
+        try {
+            // ************** UPDATE ************************
+            if (dataProvider.getId() != -1) {
+                if (!dataProviderService.hasPermissionToUpdate(params.getUser(), dataProvider.getId())) {
+                    throw new ActionDeniedException("Unauthorized user tried to update layer dataprovider - id=" + dataProvider.getId());
+                }
+                dataProviderService.update(dataProvider);
+                AuditLog.user(params.getClientIp(), params.getUser())
+                        .withParam("id", dataProvider.getId())
+                        .withParam("name", dataProvider.getName(PropertyUtil.getDefaultLanguage()))
+                        .updated(AuditLog.ResourceType.DATAPROVIDER);
+
+                ResponseHelper.writeResponse(params, dataProvider.getAsJSON());
+            }
+            // ************** INSERT ************************
+            else if (params.getUser().isAdmin()) {
+                final int id = dataProviderService.insert(dataProvider);
+                dataProvider.setId(id);
+                AuditLog.user(params.getClientIp(), params.getUser())
+                        .withParam("id", dataProvider.getId())
+                        .withParam("name", dataProvider.getName(PropertyUtil.getDefaultLanguage()))
+                        .added(AuditLog.ResourceType.DATAPROVIDER);
+
+                ResponseHelper.writeResponse(params, dataProvider.getAsJSON());
+            } else {
+                throw new ActionDeniedException("Unauthorized user tried to update layer dataprovider - id=" + dataProvider.getId());
+            }
+            flushLayerListCache();
+        } catch (Exception e) {
+            throw new ActionException("Couldn't update/insert map layer dataprovider", e);
+        }
+    }
 }
