+++ control-base/src/main/java/fi/nls/oskari/control/layer/GetMapLayerGroupsHandler.java
@@ -251,6 +251,7 @@ public class GetMapLayerGroupsHandler extends ActionHandler {
                     }
                     JSONObject json = new JSONObject();
                     JSONHelper.putValue(json, "name", provider.getName(language));
+                    JSONHelper.putValue(json, "locale", provider.getLocale());
                     // TODO: provider description will be added here in the future
                     JSONHelper.putValue(result, Integer.toString(id), json);
                 });
